Reorganized connector lifecycle module.
