+++ server/swimos_connector/src/lib.rs
@@ -26,5 +26,5 @@ pub use connector::{
 };
 pub use error::ConnectorInitError;
 pub use generic::{ConnectorAgent, MapLaneSelectorFn, ValueLaneSelectorFn};
-pub use lifecycle::ConnectorLifecycle;
+pub use lifecycle::{EgressConnectorLifecycle, IngressConnectorLifecycle};
 pub use route::ConnectorModel;
