+++ server/swimos_connector/src/lifecycle/mod.rs
@@ -12,458 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#[cfg(test)]
-mod tests;
+mod egress;
+mod ingress;
 
-use std::{cell::OnceCell, collections::HashMap};
-
-use bitflags::bitflags;
-use futures::{FutureExt, TryFutureExt};
-use swimos_agent::{
-    agent_lifecycle::{
-        item_event::{dynamic_handler, DynamicHandler, DynamicLifecycle, ItemEvent},
-        on_init::OnInit,
-        on_start::OnStart,
-        on_stop::OnStop,
-        on_timer::OnTimer,
-        HandlerContext,
-    },
-    config::{MapDownlinkConfig, SimpleDownlinkConfig},
-    event_handler::{
-        ActionContext, EventHandler, HandlerActionExt, LocalBoxHandlerAction, Sequentially,
-        TryHandlerActionExt, UnitHandler,
-    },
-    lanes::map::MapLaneEvent,
-    AgentMetadata,
-};
-use swimos_api::address::Address;
-use swimos_model::Value;
-use swimos_utilities::trigger;
-
-use crate::{
-    connector::{suspend_connector, EgressConnector, EgressConnectorSender},
-    error::ConnectorInitError,
-    ConnectorAgent, EgressContext, IngressConnector, MessageSource, SendResult,
-};
-
-/// An [agent lifecycle](swimos_agent::agent_lifecycle::AgentLifecycle) implementation that serves as an adapter for
-/// a [connector](Connector).
-#[derive(Debug, Clone, Copy)]
-pub struct ConnectorLifecycle<C>(C);
-
-impl<C> ConnectorLifecycle<C> {
-    pub fn new(connector: C) -> Self {
-        ConnectorLifecycle(connector)
-    }
-}
-
-impl<C> OnInit<ConnectorAgent> for ConnectorLifecycle<C>
-where
-    C: IngressConnector + Send,
-{
-    fn initialize(
-        &self,
-        _action_context: &mut ActionContext<ConnectorAgent>,
-        _meta: AgentMetadata,
-        _context: &ConnectorAgent,
-    ) {
-    }
-}
-
-impl<C> OnStart<ConnectorAgent> for ConnectorLifecycle<C>
-where
-    C: IngressConnector + Send,
-{
-    fn on_start(&self) -> impl EventHandler<ConnectorAgent> + '_ {
-        let ConnectorLifecycle(connector) = self;
-        let handler_context: HandlerContext<ConnectorAgent> = HandlerContext::default();
-        let (tx, rx) = trigger::trigger();
-        let suspend = handler_context
-            .effect(|| connector.create_stream())
-            .try_handler()
-            .and_then(move |stream| {
-                handler_context.suspend(async move {
-                    handler_context
-                        .value(rx.await.map_err(|_| ConnectorInitError))
-                        .try_handler()
-                        .followed_by(suspend_connector(stream))
-                })
-            });
-        connector.on_start(tx).followed_by(suspend)
-    }
-}
-
-impl<C> OnStop<ConnectorAgent> for ConnectorLifecycle<C>
-where
-    C: IngressConnector + Send,
-{
-    fn on_stop(&self) -> impl EventHandler<ConnectorAgent> + '_ {
-        self.0.on_stop()
-    }
-}
-
-impl<C> OnTimer<ConnectorAgent> for ConnectorLifecycle<C>
-where
-    C: IngressConnector + Send,
-{
-    fn on_timer(&self, _timer_id: u64) -> impl EventHandler<ConnectorAgent> + '_ {
-        UnitHandler::default()
-    }
-}
-
-impl<C> ItemEvent<ConnectorAgent> for ConnectorLifecycle<C>
-where
-    C: IngressConnector,
-{
-    type ItemEventHandler<'a> = UnitHandler
-    where
-        Self: 'a;
-
-    fn item_event<'a>(
-        &'a self,
-        _context: &ConnectorAgent,
-        _item_name: &str,
-    ) -> Option<Self::ItemEventHandler<'a>> {
-        None
-    }
-}
-
-pub struct EgressConnectorLifecycle<C: EgressConnector> {
-    lifecycle: C,
-    sender: OnceCell<C::Sender>,
-}
-
-impl<C> OnInit<ConnectorAgent> for EgressConnectorLifecycle<C>
-where
-    C: EgressConnector + Send,
-{
-    fn initialize(
-        &self,
-        _action_context: &mut ActionContext<ConnectorAgent>,
-        _meta: AgentMetadata,
-        _context: &ConnectorAgent,
-    ) {
-    }
-}
-
-impl<C> OnStop<ConnectorAgent> for EgressConnectorLifecycle<C>
-where
-    C: EgressConnector + Send,
-{
-    fn on_stop(&self) -> impl EventHandler<ConnectorAgent> + '_ {
-        UnitHandler::default()
-    }
-}
-
-bitflags! {
-
-    #[derive(Default, Debug, Copy, Clone)]
-    pub struct EgressFlags: u64 {
-        /// Initialization has completed.
-        const INITIALIZED = 0b1;
-    }
-
-}
-
-#[derive(Default, Debug)]
-struct DownlinkCollector {
-    value_downlinks: Vec<Address<String>>,
-    map_downlinks: Vec<Address<String>>,
-}
-
-impl EgressContext for DownlinkCollector {
-    fn open_value_downlink(&mut self, address: Address<String>) {
-        self.value_downlinks.push(address);
-    }
-
-    fn open_map_downlink(&mut self, address: Address<String>) {
-        self.map_downlinks.push(address);
-    }
-}
-
-impl<C> OnStart<ConnectorAgent> for EgressConnectorLifecycle<C>
-where
-    C: EgressConnector + Send,
-{
-    fn on_start(&self) -> impl EventHandler<ConnectorAgent> + '_ {
-        let EgressConnectorLifecycle { lifecycle, sender } = self;
-        let (tx, rx) = trigger::trigger();
-        let context: HandlerContext<ConnectorAgent> = Default::default();
-        let mut collector = DownlinkCollector::default();
-        let on_start = lifecycle.on_start(tx);
-        lifecycle.open_downlinks(&mut collector);
-        let create_sender = context
-            .with_parameters(|params| lifecycle.make_sender(params))
-            .try_handler()
-            .and_then(move |egress_sender| {
-                context.effect(move || {
-                    let _ = sender.set(egress_sender);
-                })
-            });
-        let await_init = context.suspend(async move {
-            context
-                .value(rx.await)
-                .try_handler()
-                .followed_by(ConnectorAgent::set_flags(EgressFlags::INITIALIZED.bits()))
-        });
-        let open_downlinks = self.open_downlinks(collector);
-        await_init
-            .followed_by(on_start)
-            .followed_by(create_sender)
-            .followed_by(open_downlinks)
-    }
-}
-
-fn is_initialized(agent: &ConnectorAgent) -> bool {
-    EgressFlags::from_bits_retain(agent.read_flags()).contains(EgressFlags::INITIALIZED)
-}
-
-impl<C> ItemEvent<ConnectorAgent> for EgressConnectorLifecycle<C>
-where
-    C: EgressConnector,
-{
-    type ItemEventHandler<'a> = DynamicHandler<'a, ConnectorAgent, Self>
-    where
-        Self: 'a;
-
-    fn item_event<'a>(
-        &'a self,
-        context: &ConnectorAgent,
-        item_name: &'a str,
-    ) -> Option<Self::ItemEventHandler<'a>> {
-        if is_initialized(context) {
-            dynamic_handler(context, self, item_name)
-        } else {
-            None
-        }
-    }
-}
-
-impl<C> DynamicLifecycle<ConnectorAgent> for EgressConnectorLifecycle<C>
-where
-    C: EgressConnector,
-{
-    type ValueHandler<'a> = LocalBoxHandlerAction<'a, ConnectorAgent, ()>
-    where
-        Self: 'a;
-
-    type MapHandler<'a> = LocalBoxHandlerAction<'a, ConnectorAgent, ()>
-    where
-        Self: 'a;
-
-    fn value_lane<'a>(
-        &'a self,
-        lane_name: &'a str,
-        _previous: Value,
-        value: &Value,
-    ) -> Self::ValueHandler<'a> {
-        let EgressConnectorLifecycle { sender, .. } = self;
-        if let Some(sender) = sender.get() {
-            let context: HandlerContext<ConnectorAgent> = Default::default();
-            handle_value_event(sender, MessageSource::Lane(lane_name), context, value)
-        } else {
-            UnitHandler::default().boxed_local()
-        }
-    }
-
-    fn map_lane<'a>(
-        &'a self,
-        lane_name: &'a str,
-        event: MapLaneEvent<Value, Value>,
-        contents: &HashMap<Value, Value>,
-    ) -> Self::MapHandler<'a> {
-        let EgressConnectorLifecycle { sender, .. } = self;
-        let kv = match &event {
-            MapLaneEvent::Clear(_) => None,
-            MapLaneEvent::Update(k, _) => {
-                let v = contents.get(k);
-                Some((k, v))
-            }
-            MapLaneEvent::Remove(k, _) => Some((k, None)),
-        };
-        if let (Some(sender), Some((k, v))) = (sender.get(), kv) {
-            let context: HandlerContext<ConnectorAgent> = Default::default();
-            handle_map_event(sender, lane_name, context, k, v)
-        } else {
-            UnitHandler::default().boxed_local()
-        }
-    }
-}
-
-impl<C> OnTimer<ConnectorAgent> for EgressConnectorLifecycle<C>
-where
-    C: EgressConnector + Send,
-{
-    fn on_timer(&self, timer_id: u64) -> impl EventHandler<ConnectorAgent> + '_ {
-        let EgressConnectorLifecycle { sender, .. } = self;
-        if let Some(sender) = sender.get() {
-            let context: HandlerContext<ConnectorAgent> = Default::default();
-            match sender.timer_event(timer_id) {
-                Some(SendResult::Suspend(f)) => {
-                    let fut = f
-                        .into_future()
-                        .map(move |h: Result<_, _>| context.value(h).try_handler().and_then(|h| h));
-                    context.suspend(fut).boxed_local()
-                }
-                Some(SendResult::RequestCallback(delay, id)) => {
-                    context.schedule_timer_event(delay, id).boxed_local()
-                }
-                _ => UnitHandler::default().boxed_local(),
-            }
-        } else {
-            UnitHandler::default().boxed_local()
-        }
-    }
-}
-
-impl<C> EgressConnectorLifecycle<C>
-where
-    C: EgressConnector,
-{
-    fn open_downlinks(
-        &self,
-        collector: DownlinkCollector,
-    ) -> impl EventHandler<ConnectorAgent> + '_ {
-        let EgressConnectorLifecycle { sender, .. } = self;
-        let context: HandlerContext<ConnectorAgent> = Default::default();
-        context
-            .value(sender.get().ok_or(ConnectorInitError))
-            .try_handler()
-            .and_then(move |sender: &C::Sender| {
-                let DownlinkCollector {
-                    value_downlinks,
-                    map_downlinks,
-                } = collector;
-                let mut open_value_dls = vec![];
-                let mut open_map_dls = vec![];
-                for address in value_downlinks {
-                    let open_dl = context
-                        .value_downlink_builder::<Value>(
-                            address.host.as_deref(),
-                            &address.node,
-                            &address.lane,
-                            SimpleDownlinkConfig::default(),
-                        )
-                        .with_shared_state((address, sender.clone()))
-                        .on_event(handle_value_event_dl)
-                        .done()
-                        .discard();
-                    open_value_dls.push(open_dl);
-                }
-                for Address { host, node, lane } in map_downlinks {
-                    let open_dl = context
-                        .map_downlink_builder::<Value, Value>(
-                            host.as_deref(),
-                            &node,
-                            &lane,
-                            MapDownlinkConfig::default(),
-                        )
-                        .with_state(("dl".to_string(), sender.clone()))
-                        .on_update(handle_update)
-                        .on_remove(handle_remove)
-                        .done()
-                        .discard();
-
-                    open_map_dls.push(open_dl);
-                }
-                Sequentially::new(open_value_dls).followed_by(Sequentially::new(open_map_dls))
-            })
-    }
-}
-
-fn handle_value_event_dl<'a, S, E>(
-    state: &'a (Address<String>, S),
-    context: HandlerContext<ConnectorAgent>,
-    value: &Value,
-) -> impl EventHandler<ConnectorAgent> + 'a
-where
-    S: EgressConnectorSender<E>,
-    E: std::error::Error + Send + 'static,
-{
-    let (addr, sender) = state;
-    handle_value_event(sender, MessageSource::Downlink(addr), context, value)
-}
-
-fn handle_update<'a, S, E>(
-    state: &'a (String, S),
-    context: HandlerContext<ConnectorAgent>,
-    _map: &HashMap<Value, Value>,
-    key: Value,
-    _previous: Option<Value>,
-    new_value: &Value,
-) -> impl EventHandler<ConnectorAgent> + 'a
-where
-    S: EgressConnectorSender<E>,
-    E: std::error::Error + Send + 'static,
-{
-    let (name, sender) = state;
-    handle_map_event(sender, name, context, &key, Some(new_value))
-}
-
-fn handle_remove<'a, S, E>(
-    state: &'a (String, S),
-    context: HandlerContext<ConnectorAgent>,
-    _map: &HashMap<Value, Value>,
-    key: Value,
-    _previous: Value,
-) -> impl EventHandler<ConnectorAgent> + 'a
-where
-    S: EgressConnectorSender<E>,
-    E: std::error::Error + Send + 'static,
-{
-    let (name, sender) = state;
-    handle_map_event(sender, name, context, &key, None)
-}
-fn handle_map_event<'a, S, E>(
-    sender: &'a S,
-    name: &'a str,
-    context: HandlerContext<ConnectorAgent>,
-    key: &Value,
-    value: Option<&Value>,
-) -> LocalBoxHandlerAction<'a, ConnectorAgent, ()>
-where
-    S: EgressConnectorSender<E>,
-    E: std::error::Error + Send + 'static,
-{
-    match sender.send(
-        MessageSource::Lane(name),
-        Some(key),
-        value.unwrap_or(&Value::Extant),
-    ) {
-        Some(SendResult::Suspend(f)) => {
-            let fut = f
-                .into_future()
-                .map(move |h: Result<_, _>| context.value(h).try_handler().and_then(|h| h));
-            context.suspend(fut).boxed_local()
-        }
-        Some(SendResult::RequestCallback(delay, id)) => {
-            context.schedule_timer_event(delay, id).boxed_local()
-        }
-        _ => UnitHandler::default().boxed_local(),
-    }
-}
-
-fn handle_value_event<'a, S, E>(
-    sender: &'a S,
-    source: MessageSource<'_>,
-    context: HandlerContext<ConnectorAgent>,
-    value: &Value,
-) -> LocalBoxHandlerAction<'a, ConnectorAgent, ()>
-where
-    S: EgressConnectorSender<E>,
-    E: std::error::Error + Send + 'static,
-{
-    match sender.send(source, None, value) {
-        Some(SendResult::Suspend(f)) => {
-            let fut = f
-                .into_future()
-                .map(move |h: Result<_, _>| context.value(h).try_handler().and_then(|h| h));
-            context.suspend(fut).boxed_local()
-        }
-        Some(SendResult::RequestCallback(delay, id)) => {
-            context.schedule_timer_event(delay, id).boxed_local()
-        }
-        Some(SendResult::Fail(err)) => context.fail(err).boxed_local(),
-        _ => UnitHandler::default().boxed_local(),
-    }
-}
+pub use egress::EgressConnectorLifecycle;
+pub use ingress::IngressConnectorLifecycle;
